requirements_conda.txt
numpy
opencv
pandas
Pillow
scikit-image
scipy==1.2.0
pytorch
torchvision
nltk